refactor: divvy up tokenserver-postgres's db module
Closes STOR-405
